Format README.md
I feel a bit rude doing this, but I really cannot work with plaintext
that's not wrapped to 80 characters (maybe 100 if pushed) in any
editor I use. I'm planning to add some more resources about acme to
make it friendly to new users, and if I'm to do that, I want the
README to be pleasant to work with.